Move cities list generation to a separate component
