+++ posleurokov-app/src/components/Cities/Cities.module.scss
@@ -1,29 +1,21 @@
 @import 'src/assets/styles/_variables.scss';
 
-.cities {
+.all_cities {
   margin-top: 19px;
   display: flex;
-  text-decoration: none;
   color: #5F6060;
-  padding-bottom: 12px;
-  border-bottom: 1px solid #EEEFEF;
   font-size: 14px;
-  font-weight: 400;
   line-height: 16px;
+  font-weight: 700;
+  padding-bottom: 0%;
+  text-decoration: underline;
+  text-decoration-style: dashed;
+  text-underline-offset: 3px;
 
   img {
     margin-top: auto;
     margin-bottom: auto;
     margin-right: 10px;
-  }
-
-  &:last-child {
-    font-weight: 700;
-    padding-bottom: 0%;
-    text-decoration: underline;
-    text-decoration-style: dashed;
-    text-underline-offset: 3px;
-    border-bottom: none;
   }
 }
